Separate main.py into main.py and update.py
